# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A web service similar to httpbin.org built with Fiber.",
        "title": "Fiber HTTPBin API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/delay": {
            "get": {
                "description": "Delays the response by a specified number of seconds",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Delay"
                ],
                "summary": "Delays the response",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Delay in seconds",
                        "name": "seconds",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/post": {
            "post": {
                "description": "Echoes the POST request data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "POST"
                ],
                "summary": "Returns the POST data",
                "parameters": [
                    {
                        "description": "POST Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/status": {
            "get": {
                "description": "Returns a response with the given HTTP status code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Status"
                ],
                "summary": "Returns specified HTTP status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "HTTP Status Code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
...
}